Updating: Tests structure and documentation
