+++ pyproject.toml
@@ -8,6 +8,12 @@ build-backend = "setuptools.build_meta"
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
+markers = [
+  "unit: Fast tests that isolate a small unit of behavior with controlled doubles.",
+  "integration: Tests that exercise the library against real external behavior or package boundaries.",
+  "validation: Algorithmic equivalence and behavior-validation tests across multiple k values.",
+  "slow: Tests that are intentionally heavier and may take longer to execute.",
+]
 
 [tool.coverage.run]
 branch = true
@@ -27,4 +33,5 @@ ignore = ["E501"]
 
 [tool.ruff.lint.per-file-ignores]
 "tests/conftest.py" = ["E402"]
-"tests/test_reference_equivalence.py" = ["E402"]
+"tests/integration/test_reference_equivalence.py" = ["E402"]
+"tests/validation/conftest.py" = ["E402"]
