+++ tests/validation/__init__.py
@@ -1,0 +1,1 @@
+"""Validation tests for Core-SG."""
